ci: Remove `extensions::run_tests` workflow (#14)
This PR removes the `run_tests.yml` workflow file, as it is no longer
needed after moving the organization into the Zed GitHub enterprise.

We now run the tests on pull requests with the workflow defined at
https://github.com/zed-extensions/workflows and invoke this using an
organization-wide policy for `zed-extensions`.

Aside from the duplicate testing, having both files causes issues with
concurrency, hence the removal of that file here.